refactor(status): computeAvgLatency gets its own file
Move the bucket-array `computeAvgLatency` out of `computeUptime.ts` into
its own module with colocated tests. Split `computeUptime.test.ts` along
the same boundary. The services-level variant in `summaryStats.ts` stays
put — different shape, different call site.
